Web/requirements.txt
flask~=2.2.3
requests
python-dotenv